@@ -1,40 +1,40 @@
 {
 	"name": "bili-sync-web",
-	"version": "2.9.3",
+	"version": "2.10.2",
 	"devDependencies": {
-		"@eslint/compat": "^1.2.5",
-		"@eslint/js": "^9.18.0",
-		"@internationalized/date": "^3.8.1",
+		"@eslint/compat": "^1.4.1",
+		"@eslint/js": "^9.39.2",
+		"@internationalized/date": "^3.10.1",
 		"@lucide/svelte": "^0.544.0",
-		"@sveltejs/adapter-static": "^3.0.8",
-		"@sveltejs/kit": "2.22.2",
-		"@sveltejs/vite-plugin-svelte": "^6.0.0",
-		"@tailwindcss/forms": "^0.5.9",
-		"@tailwindcss/typography": "^0.5.15",
-		"@tailwindcss/vite": "^4.0.0",
+		"@sveltejs/adapter-static": "^3.0.10",
+		"@sveltejs/kit": "^2.49.2",
+		"@sveltejs/vite-plugin-svelte": "^6.2.1",
+		"@tailwindcss/forms": "^0.5.11",
+		"@tailwindcss/typography": "^0.5.19",
+		"@tailwindcss/vite": "^4.1.18",
 		"@types/d3-scale": "^4.0.9",
 		"@types/d3-shape": "^3.1.7",
-		"bits-ui": "^2.11.0",
+		"bits-ui": "^2.15.2",
 		"clsx": "^2.1.1",
-		"eslint": "^9.18.0",
-		"eslint-config-prettier": "^10.0.1",
-		"eslint-plugin-svelte": "^3.0.0",
-		"globals": "^16.0.0",
-		"layerchart": "2.0.0-next.27",
+		"eslint": "^9.39.2",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.13.1",
+		"globals": "^16.5.0",
+		"layerchart": "^2.0.0-next.43",
 		"mode-watcher": "^1.1.0",
-		"prettier": "^3.4.2",
-		"prettier-plugin-svelte": "^3.3.3",
-		"prettier-plugin-tailwindcss": "^0.6.11",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"svelte-sonner": "^1.0.1",
-		"tailwind-merge": "^3.0.2",
+		"prettier": "^3.7.4",
+		"prettier-plugin-svelte": "^3.4.1",
+		"prettier-plugin-tailwindcss": "^0.7.2",
+		"svelte": "^5.46.1",
+		"svelte-check": "^4.3.5",
+		"svelte-sonner": "^1.0.7",
+		"tailwind-merge": "^3.4.0",
 		"tailwind-variants": "^1.0.0",
-		"tailwindcss": "^4.0.0",
-		"tw-animate-css": "^1.3.2",
-		"typescript": "^5.0.0",
-		"typescript-eslint": "^8.20.0",
-		"vite": "7.0.3"
+		"tailwindcss": "^4.1.18",
+		"tw-animate-css": "^1.4.0",
+		"typescript": "^5.9.3",
+		"typescript-eslint": "^8.51.0",
+		"vite": "^7.3.0"
 	},
 	"private": true,
 	"scripts": {
@@ -47,5 +47,9 @@
 		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint ."
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"@types/qrcode": "^1.5.6",
+		"qrcode": "^1.5.4"
+	}
 }
