@@ -1,9 +1,5 @@
 use rand::seq::IndexedRandom;
 
-pub(super) fn default_time_format() -> String {
-    "%Y-%m-%d".to_string()
-}
-
 /// 默认的 auth_token 实现，生成随机 16 位字符串
 pub(super) fn default_auth_token() -> String {
     let byte_choices = b"abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()_+-=";
@@ -13,6 +9,22 @@ pub(super) fn default_auth_token() -> String {
         .collect()
 }
 
-pub(super) fn default_bind_address() -> String {
+pub(crate) fn default_bind_address() -> String {
     "0.0.0.0:12345".to_string()
+}
+
+pub(super) fn default_time_format() -> String {
+    "%Y-%m-%d".to_string()
+}
+
+pub fn default_favorite_path() -> String {
+    "收藏夹/{{name}}".to_owned()
+}
+
+pub fn default_collection_path() -> String {
+    "合集/{{name}}".to_owned()
+}
+
+pub fn default_submission_path() -> String {
+    "投稿/{{name}}".to_owned()
 }
