@@ -1,5 +1,7 @@
 import { writable } from 'svelte/store';
 
+export type StatusFilterValue = 'failed' | 'succeeded' | 'waiting' | null;
+
 export interface AppState {
 	query: string;
 	currentPage: number;
@@ -7,25 +9,30 @@ export interface AppState {
 		type: string;
 		id: string;
 	} | null;
+	statusFilter: StatusFilterValue | null;
 }
 
 export const appStateStore = writable<AppState>({
 	query: '',
 	currentPage: 0,
-	videoSource: null
+	videoSource: null,
+	statusFilter: null
 });
 
 export const ToQuery = (state: AppState): string => {
-	const { query, videoSource } = state;
+	const { query, videoSource, currentPage, statusFilter } = state;
 	const params = new URLSearchParams();
-	if (state.currentPage > 0) {
-		params.set('page', String(state.currentPage));
+	if (currentPage > 0) {
+		params.set('page', String(currentPage));
 	}
 	if (query.trim()) {
 		params.set('query', query);
 	}
 	if (videoSource && videoSource.type && videoSource.id) {
 		params.set(videoSource.type, videoSource.id);
+	}
+	if (statusFilter) {
+		params.set('status_filter', statusFilter);
 	}
 	const queryString = params.toString();
 	return queryString ? `videos?${queryString}` : 'videos';
@@ -40,6 +47,7 @@ export const ToFilterParams = (
 	favorite?: number;
 	submission?: number;
 	watch_later?: number;
+	status_filter?: Exclude<StatusFilterValue, null>;
 } => {
 	const params: {
 		query?: string;
@@ -47,6 +55,7 @@ export const ToFilterParams = (
 		favorite?: number;
 		submission?: number;
 		watch_later?: number;
+		status_filter?: Exclude<StatusFilterValue, null>;
 	} = {};
 
 	if (state.query.trim()) {
@@ -57,13 +66,15 @@ export const ToFilterParams = (
 		const { type, id } = state.videoSource;
 		params[type as 'collection' | 'favorite' | 'submission' | 'watch_later'] = parseInt(id);
 	}
-
+	if (state.statusFilter) {
+		params.status_filter = state.statusFilter;
+	}
 	return params;
 };
 
 // 检查是否有活动的筛选条件
 export const hasActiveFilters = (state: AppState): boolean => {
-	return !!(state.query.trim() || state.videoSource);
+	return !!(state.query.trim() || state.videoSource || state.statusFilter);
 };
 
 export const setQuery = (query: string) => {
@@ -73,24 +84,17 @@ export const setQuery = (query: string) => {
 	}));
 };
 
-export const setVideoSourceFilter = (filter: { type: string; id: string }) => {
-	appStateStore.update((state) => ({
-		...state,
-		videoSource: filter
-	}));
-};
-
-export const clearVideoSourceFilter = () => {
-	appStateStore.update((state) => ({
-		...state,
-		videoSource: null
-	}));
-};
-
 export const setCurrentPage = (page: number) => {
 	appStateStore.update((state) => ({
 		...state,
 		currentPage: page
+	}));
+};
+
+export const setStatusFilter = (statusFilter: StatusFilterValue | null) => {
+	appStateStore.update((state) => ({
+		...state,
+		statusFilter
 	}));
 };
 
@@ -104,19 +108,13 @@ export const resetCurrentPage = () => {
 export const setAll = (
 	query: string,
 	currentPage: number,
-	videoSource: { type: string; id: string } | null
+	videoSource: { type: string; id: string } | null,
+	statusFilter: StatusFilterValue | null
 ) => {
 	appStateStore.set({
 		query,
 		currentPage,
-		videoSource
-	});
-};
-
-export const clearAll = () => {
-	appStateStore.set({
-		query: '',
-		currentPage: 0,
-		videoSource: null
+		videoSource,
+		statusFilter
 	});
 };
